requirements.txt
upstox-python-sdk>=2.21.0
pyarrow>=15.0.0
pandas>=2.2.0
numpy>=1.26.0
websockets>=12.0
protobuf>=5.26.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas-ta>=0.3.14b
httpx>=0.27.0
openai>=1.30.0
anthropic>=0.25.0
